pyproject.toml
[tool.poetry]
name = "cassandradb"
version = "0.1.0"
description = ""
authors = ["Sambiase <marsamb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cassandra-driver = "^3.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
